+++ src/main/java/it/gov/pagopa/pu/organization/dto/orgsilservice/SilServiceAuthConfigDTO.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.pu.organization.model;
+package it.gov.pagopa.pu.organization.dto.orgsilservice;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
