+++ src/main/java/it/gov/pagopa/pu/organization/repository/OrgSilServiceRepository.java
@@ -1,7 +1,7 @@
 package it.gov.pagopa.pu.organization.repository;
 
 import it.gov.pagopa.pu.organization.enums.OrgSilServiceType;
-import it.gov.pagopa.pu.organization.model.OrgSilService;
+import it.gov.pagopa.pu.organization.model.orgsilservice.OrgSilService;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
